pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hiceebox"
version = "0.1.0"
description = "Hi-C + multi-omics visualization toolkit for genomic data"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["genomics", "hi-c", "visualization", "bioinformatics", "multi-omics"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.20",
    "matplotlib>=3.5",
    "pyyaml>=5.4",
]

[project.optional-dependencies]
hic = ["hic-straw>=1.3"]
mcool = ["cooler>=0.9"]
bigwig = ["pyBigWig>=0.3"]
gui = [
    "PySide6>=6.5",
    "hic-straw>=1.3",
    "cooler>=0.9",
    "pyBigWig>=0.3",
    "pandas>=1.5",
]
all = [
    "hic-straw>=1.3",
    "cooler>=0.9",
    "pyBigWig>=0.3",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
]

[project.scripts]
hiceebox = "hiceebox.cli.main:main"
hiceebox-gui = "hiceebox_gui.app:main"

[project.urls]
Homepage = "https://github.com/yourusername/hiceebox"
Documentation = "https://hiceebox.readthedocs.io"
Repository = "https://github.com/yourusername/hiceebox"
Issues = "https://github.com/yourusername/hiceebox/issues"

[tool.setuptools]
packages = [
    "hiceebox",
    "hiceebox.matrix",
    "hiceebox.tracks",
    "hiceebox.view",
    "hiceebox.cli",
    "hiceebox.utils",
    "hiceebox_gui"
]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Briefcase configuration for macOS packaging
[tool.briefcase]
project_name = "HiCeeBox"
bundle = "com.hiceebox"

[tool.briefcase.app.hiceebox]
formal_name = "HiCeeBox"
description = "Interactive genome browser for Hi-C and multi-omics data"
long_description = """HiCeeBox is a desktop application for interactive browsing of Hi-C contact maps and genomic annotations. Features include real-time navigation, multi-track visualization, and high-quality export."""
icon = "assets/icon"
sources = ["hiceebox", "hiceebox_gui"]
requires = [
    "numpy>=1.20",
    "matplotlib>=3.5",
    "pyyaml>=5.4",
    "PySide6>=6.5",
    "hic-straw>=1.3",
    "cooler>=0.9",
    "pyBigWig>=0.3",
    "pandas>=1.5",
]

[tool.briefcase.app.hiceebox.macOS]
requires = []
universal_build = true

[tool.briefcase.app.hiceebox.macOS.app_packages]
# Additional packages for macOS

